[stm32/port] resolve stack related problems
Issue still left to fix:
1. We shouldl not maintain two copy of porting code on each plat.
2. There is only one ousia_cfg.h necessary.

+++ platform/stm32/port/ousia_port.h
@@ -62,7 +62,7 @@ void _port_assert_fail(const char *file, int line, const char *exp);
 void _port_init_printf(void **stdout_putp, void (**stdout_putf)(void *dev, char ch));
 void _port_systick_init(void (*callback)(void));
 void _port_context_switch(void *curr_pcb, void *target_pcb) __attribute__ ((naked));
-uint8 *_port_process_stack_init(void *pentry, void *args, void *stack_base);
+uint32 *_port_process_stack_init(void *pentry, void *args, void *stack_base);
 
 
 #endif /* OUSIA_PORT_H */
